+++ README.md
@@ -14,7 +14,7 @@
 - [x] 환율 정보 로딩중임을 사용자가 알 수 있도록 UI를 추가해 주세요.
 - [x] 금액은 세자리 마다 콤마(,) 를 표시하세요
 - [x] 유로화는 소수점 2자리까지 표시하고, 원화는 소수점을 표시하지 않습니다.
-- [x] 비니지스 로직을 분리하고 MVVM 구현패턴을 적용하세요.
+- [x] 비즈니스 로직을 분리하고 MVVM 구현패턴을 적용하세요.
 
 #### 옵션 수행 항목
