+++ src/types/Euroinfo.ts
@@ -1,0 +1,9 @@
+export interface EuroInfo {
+  basePrice: number;
+  openingPrice: number;
+  changePrice: number;
+  cashBuyingPrice: number;
+  cashSellingPrice: number;
+  ttSellingPrice: number;
+  ttBuyingPrice: number;
+}
